more bug fixes

+++ src/main/java/frc/robot/subsystems/MechanicalParts/lilElevatorConveyerBeltThingy.java
@@ -3,10 +3,16 @@ package frc.robot.subsystems.MechanicalParts;
 import com.ctre.phoenix.motorcontrol.FeedbackDevice;
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonFX;
+import com.revrobotics.CANSparkMax;
+import com.revrobotics.RelativeEncoder;
+import com.revrobotics.SparkMaxPIDController;
+
+import edu.wpi.first.wpilibj.I2C;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 
 import java.lang.Math;
+import java.nio.ByteBuffer;
 
 //STRATEGY
 //CONSTANTLY MOVE SLOWLY UNTILL COLOR SENSOR DETECTION, THEN WHEN SHOOT COMMAND MOVE FORWARDS FOR A SECOND
@@ -51,7 +57,7 @@ public class lilElevatorConveyerBeltThingy extends SubsystemBase {
     }
 
     public boolean IsCubeDetected(){
-        int closeness = Math.abs(Constants.ElevatorConveyerThing.r - getRed()) + Math.abs(Constants.ElevatorConveyerThing.g - getGreen()) + Math.abs(Constants.ElevatorConveyerThing.b - getBlue());
+        int closeness = (int) (Math.abs(Constants.ElevatorConveyerThing.r - getRed()) + Math.abs(Constants.ElevatorConveyerThing.g - getGreen()) + Math.abs(Constants.ElevatorConveyerThing.b - getBlue()));
         return (closeness < Constants.ElevatorConveyerThing.WiggleRoom);
     }
